Fixed debugger to properly display u16::MAX address

+++ src/core/gb.rs
@@ -72,9 +72,9 @@ impl Gameboy {
         self.cpu.reg.pc
     }
 
-    /// Returns a boxed slice of u8 values contained within the given range of u16 values.
+    /// Returns a boxed slice of u8 values contained within the given range of usize values.
     /// Only returns values as read via the CPU, so forbidden or fixed reads will not be bypassed
-    pub fn get_memory_range(&self, range: std::ops::Range<u16>) -> Box<[u8]> {
+    pub fn get_memory_range(&self, range: std::ops::Range<usize>) -> Box<[u8]> {
         self.mmu.get_memory_range(range).into_boxed_slice()
     }
 }
